+++ snuba/datasets/storages/generic_metrics.py
@@ -3,18 +3,7 @@ The storages defined in this file are for the generic metrics system,
 initially built to handle metrics-enhanced performance.
 """
 
-import collections
-from dataclasses import dataclass, fields
-from typing import (
-    Any,
-    Callable,
-    Mapping,
-    MutableSequence,
-    Optional,
-    Sequence,
-    Type,
-    Union,
-)
+from typing import Sequence
 
 from arroyo import Topic as KafkaTopic
 from arroyo.backends.kafka import KafkaProducer
@@ -38,6 +27,15 @@ from snuba.clickhouse.columns import (
 )
 from snuba.clickhouse.processors import QueryProcessor
 from snuba.clusters.storage_sets import StorageSetKey
+from snuba.datasets.configurations.config_loader import (
+    CONF_TO_PREFILTER,
+    CONF_TO_PROCESSOR,
+    StorageConfig,
+    dataclass_from_dict,
+    deep_compare_storages,
+    parse_columns,
+    policy_creator_creator,
+)
 from snuba.datasets.generic_metrics_processor import (
     GenericDistributionsMetricsProcessor,
     GenericSetsMetricsProcessor,
@@ -47,9 +45,6 @@ from snuba.datasets.metrics_messages import InputType
 from snuba.datasets.schemas.tables import TableSchema, WritableTableSchema
 from snuba.datasets.storage import ReadableTableStorage, WritableTableStorage
 from snuba.datasets.storages import StorageKey
-from snuba.datasets.storages.deep_compare import deep_compare_storages
-
-# from snuba.datasets.storages.deep_compare import deep_compare_storages
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
 from snuba.query.processors.table_rate_limit import TableRateLimit
 from snuba.query.processors.tuple_unaliaser import TupleUnaliaser
@@ -199,164 +194,11 @@ distributions_bucket_storage_old = WritableTableStorage(
 )
 
 
-def dataclass_from_dict(cls: Type[Any], d: Any) -> Any:
-    # recursively type cast dict to classes
-    # HACK: super hacks
-    try:
-        if cls.__origin__ == collections.abc.Sequence:
-            return [dataclass_from_dict(cls.__dict__["__args__"][0], cfg) for cfg in d]
-        elif cls.__origin__ == Union:
-            if isinstance(d[0], str) or isinstance(d[0], int):
-                return d
-            elif isinstance(d[0], dict):
-                for t in cls.__dict__["__args__"]:
-                    result = dataclass_from_dict(t, d)
-                    if not isinstance(result[0], dict):
-                        return result
-            return d
-    except Exception:
-        pass
-
-    if cls in {str, int}:
-        return d
-
-    # https://stackoverflow.com/a/54769644
-    try:
-        fieldtypes = {f.name: f.type for f in fields(cls)}
-        return cls(**{f: dataclass_from_dict(fieldtypes[f], d[f]) for f in d})
-    except Exception:
-        return d
-
-
-@dataclass(frozen=True)
-class StorageMetadataConfig:
-    key: str
-    set_key: str
-
-
-@dataclass(frozen=True)
-class NestedColumnConfig:
-    name: str
-    type: str
-    args: Sequence[int]
-
-
-@dataclass(frozen=True)
-class ColumnConfig:
-    name: str
-    type: str
-    args: Union[Sequence[NestedColumnConfig], Sequence[Union[str, int]]]
-
-
-@dataclass(frozen=True)
-class SchemaConfig:
-    # columns: Sequence[Union[Column[SchemaModifiers], tuple[str, ColumnType[SchemaModifiers]]]]
-    # somehow configure this^
-    columns: Sequence[ColumnConfig]
-    local_table_name: str
-    dist_table_name: str
-
-
-@dataclass(frozen=True)
-class FunctionCallConfig:
-    type: str
-    args: Sequence[str]
-
-
-@dataclass(frozen=True)
-class StreamLoaderConfig:
-    processor: str
-    default_topic: str
-    commit_log_topic: str
-    subscription_scheduled_topic: str
-    subscription_scheduler_mode: str
-    subscription_result_topic: str
-    replacement_topic: Optional[str]
-    pre_filter: FunctionCallConfig
-    dlq_policy: FunctionCallConfig
-
-
-@dataclass(frozen=True)
-class StorageConfig:
-    storage: StorageMetadataConfig
-    schema: SchemaConfig
-    stream_loader: StreamLoaderConfig
-
-
-def policy_creator_creator(
-    dlq_policy_conf: FunctionCallConfig,
-) -> Optional[Callable[[], DeadLetterQueuePolicy]]:
-    if dlq_policy_conf.type == "produce":
-        dlq_topic = dlq_policy_conf.args[0]
-
-        def produce_policy_creator2() -> DeadLetterQueuePolicy:
-            return ProduceInvalidMessagePolicy(
-                KafkaProducer(build_kafka_producer_configuration(Topic(dlq_topic))),
-                KafkaTopic(dlq_topic),
-            )
-
-        return produce_policy_creator2
-    return None
-
-
 file = open("./snuba/datasets/configurations/generic_metrics.yaml")
 conf_yml = safe_load(file)
 assert isinstance(conf_yml, dict)
 conf = dataclass_from_dict(StorageConfig, conf_yml)
 assert isinstance(conf, StorageConfig)
-
-# print(conf)
-
-CONF_TO_PREFILTER: Mapping[str, Any] = {
-    "kafka_header_select_filter": KafkaHeaderSelectFilter
-}
-CONF_TO_PROCESSOR: Mapping[str, Any] = {
-    "generic_distributions_metrics_processor": GenericDistributionsMetricsProcessor
-}
-
-
-def parse_simple(
-    col: Union[ColumnConfig, NestedColumnConfig]
-) -> Column[SchemaModifiers]:
-    if col.type == "UInt":
-        assert isinstance(col.args[0], int)
-        return Column(col.name, UInt(col.args[0]))
-    elif col.type == "Float":
-        assert isinstance(col.args[0], int)
-        return Column(col.name, Float(col.args[0]))
-    elif col.type == "String":
-        return Column(col.name, String())
-    elif col.type == "DateTime":
-        return Column(col.name, DateTime())
-    raise
-
-
-def parse_columns(
-    columns: Union[Sequence[ColumnConfig], Sequence[NestedColumnConfig]]
-) -> Sequence[Column[SchemaModifiers]]:
-    cols: MutableSequence[Column[SchemaModifiers]] = []
-
-    SIMPLE_COLUMN_TYPES = {
-        "UInt": UInt,
-        "Float": Float,
-        "String": String,
-        "DateTime": DateTime,
-    }
-
-    for col in columns:
-        column: Optional[Column[SchemaModifiers]] = None
-        if col.type in SIMPLE_COLUMN_TYPES:
-            column = parse_simple(col)
-        elif col.type == "Nested":
-            column = Column(col.name, Nested(parse_columns(col.args)))  # type: ignore
-        elif col.type == "Array":
-            subtype, value = col.args
-            assert isinstance(subtype, str)
-            assert isinstance(value, int)
-            column = Column(col.name, Array(SIMPLE_COLUMN_TYPES[subtype](value)))
-        assert column is not None
-        cols.append(column)
-    return cols
 
 
 distributions_bucket_storage = WritableTableStorage(
